requirements.txt
python-dotenv==1.0.1
google-generativeai==0.3.2
streamlit==1.30.0
langdetect
googletrans==4.0.0-rc1
deep-translator==1.4.2